refactor(settings): 云端模型目录与提供商节点代码归位去重
- 前端：自 SettingsView.jsx 抽出 model-catalog.js（云端模型目录、预设模板、
  品牌图标与 provider 匹配纯函数），目录相关三语文案补丁移入 settings-i18n.js
- Rust：core/model_endpoint.rs 归位 Anthropic 端点判定、OpenAI/Ollama/
  LM Studio 模型探测（自 monitor/settings 收敛）与 review/memory 的
  Messages 直连（自 #135 合入）；probe_openai_models 统一走 models_probe_url
- platform/prefs.rs 拆分为 prefs/mod.rs + prefs/model.rs；ModelPreset 收敛
  context_window_fallback、default_model/base_url，memory/review/monitor/bridge
  删除重复实现改为委托；ensure_direct_chat_compatible 随 #135 Messages
  直连落地移除（Anthropic 不再前置报错）
