+++ pinvou3-app/src-tauri/src/app/commands/monitor.rs
@@ -68,9 +68,9 @@ pub async fn discover_local_vllm(
         // 按框架选探测方式：Ollama / LM Studio 的列表接口返回全部已下载模型，
         // 需用各自原生接口区分已加载；vLLM 等 served 即已加载。
         let probe = match local_port_of(&base_url) {
-            Some(11434) => crate::features::monitor::probe_ollama_models(&base_url).await,
-            Some(1234) => crate::features::monitor::probe_lmstudio_models(&base_url).await,
-            _ => crate::features::monitor::probe_openai_models(&base_url)
+            Some(11434) => crate::core::model_endpoint::probe_ollama_models(&base_url).await,
+            Some(1234) => crate::core::model_endpoint::probe_lmstudio_models(&base_url).await,
+            _ => crate::core::model_endpoint::probe_openai_models(&base_url)
                 .await
                 .map(|mut probe| {
                     for model in &mut probe.models {
