+++ pinvou3-app/src-tauri/src/features/monitor/mod.rs
@@ -16,6 +16,7 @@ use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 use parking_lot::Mutex;
 use serde::Serialize;
 
+use crate::core::model_endpoint::{is_anthropic_endpoint, models_probe_url, strip_v1_suffix};
 use crate::platform::credential_store::{CredentialStore, SystemCredentialStore};
 use crate::platform::prefs::{ModelPreset, SavedModel, UserPrefs};
 
@@ -837,241 +838,20 @@ fn base_model_snapshot(
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OpenAiModelInfo {
-    pub id: String,
-    pub max_model_len: Option<u32>,
-    /// 是否已加载到内存。`None` = 未知（通用 OpenAI 兼容端点不区分）。
-    /// Ollama（/api/ps vs /api/tags）与 LM Studio（/api/v0/models 的 state）
-    /// 的列表接口返回全部已下载模型，二者都是 JIT 加载——任何推理请求引用
-    /// 模型名就会静默载入内存。探测必须把这个状态传给前端，避免把未加载的
-    /// 大模型当作"就绪"自动填充。
-    pub loaded: Option<bool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OpenAiModelsProbe {
-    pub models: Vec<OpenAiModelInfo>,
-}
-
 fn parse_models_response(v: serde_json::Value) -> Option<(Option<String>, Option<u32>)> {
-    let first = parse_models_response_list(v)?.into_iter().next()?;
+    let first = crate::core::model_endpoint::parse_models_response_list(v)?
+        .into_iter()
+        .next()?;
     Some((Some(first.id), first.max_model_len))
-}
-
-fn parse_models_response_list(v: serde_json::Value) -> Option<Vec<OpenAiModelInfo>> {
-    let data = v.get("data")?.as_array()?;
-    let models = data
-        .iter()
-        .filter_map(|item| {
-            let id = item.get("id").and_then(|v| v.as_str())?.trim();
-            if id.is_empty() {
-                return None;
-            }
-            let max_model_len = item
-                .get("max_model_len")
-                .and_then(|v| v.as_u64())
-                .map(|n| n as u32);
-            Some(OpenAiModelInfo {
-                id: id.to_string(),
-                max_model_len,
-                loaded: None,
-            })
-        })
-        .collect::<Vec<_>>();
-    (!models.is_empty()).then_some(models)
-}
-
-/// 通用 OpenAI 兼容 `/models` 探测。探测地址与云端 probe / 连接测试同一口径
-/// （`models_probe_url`）：upstream 不带 `/v1` 也不补——glm `/paas/v4`、火山方舟
-/// `/api/v3`、gemini `/v1beta/openai` 的 `/models` 端点均存在，补 `/v1` 会拼成
-/// 不存在的地址永远 404。本地候选（vLLM/Ollama/LM Studio）由 discover 统一
-/// 归一成 `/v1` 结尾后传入，行为不变。
-pub async fn probe_openai_models(base_url: &str) -> Option<OpenAiModelsProbe> {
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(3))
-        .build()
-        .ok()?;
-    let url = models_probe_url(base_url);
-    let resp = client.get(url).send().await.ok()?;
-    if !resp.status().is_success() {
-        return None;
-    }
-    let v = resp.json::<serde_json::Value>().await.ok()?;
-    Some(OpenAiModelsProbe {
-        models: parse_models_response_list(v)?,
-    })
-}
-
-/// Ollama `/api/ps` 返回的已加载模型名集合。解析失败按空集处理
-/// （宁可全部标未加载，也不错标已加载）。
-fn parse_ollama_ps_names(v: serde_json::Value) -> std::collections::HashSet<String> {
-    v.get("models")
-        .and_then(|m| m.as_array())
-        .map(|models| {
-            models
-                .iter()
-                .filter_map(|item| {
-                    item.get("name")
-                        .or_else(|| item.get("model"))
-                        .and_then(|v| v.as_str())
-                        .map(|s| s.trim().to_string())
-                })
-                .filter(|s| !s.is_empty())
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-/// Ollama `/api/tags` 返回的已下载模型名列表（保持顺序、去重）。
-fn parse_ollama_tag_names(v: serde_json::Value) -> Vec<String> {
-    let mut out: Vec<String> = Vec::new();
-    if let Some(models) = v.get("models").and_then(|m| m.as_array()) {
-        for item in models {
-            let Some(name) = item.get("name").and_then(|v| v.as_str()) else {
-                continue;
-            };
-            let name = name.trim();
-            if !name.is_empty() && !out.iter().any(|existing| existing == name) {
-                out.push(name.to_string());
-            }
-        }
-    }
-    out
-}
-
-/// LM Studio 原生 REST `/api/v0/models`：每项带 `state`（loaded / not-loaded）。
-/// 返回 `None` 表示响应形状不认识，调用方回退 OpenAI 兼容探测。
-fn parse_lmstudio_v0_models(v: &serde_json::Value) -> Option<Vec<OpenAiModelInfo>> {
-    let data = v.get("data")?.as_array()?;
-    let models = data
-        .iter()
-        .filter_map(|item| {
-            let id = item.get("id").and_then(|v| v.as_str())?.trim();
-            if id.is_empty() {
-                return None;
-            }
-            let loaded = item
-                .get("state")
-                .and_then(|v| v.as_str())
-                .map(|state| state == "loaded");
-            Some(OpenAiModelInfo {
-                id: id.to_string(),
-                max_model_len: None,
-                loaded,
-            })
-        })
-        .collect::<Vec<_>>();
-    (!models.is_empty()).then_some(models)
-}
-
-/// 探测 Ollama：区分"已加载"（/api/ps）与"仅下载未加载"（/api/tags）。
-/// 两个接口都是只读列表，不会触发加载；绝不能用推理请求探测。
-pub async fn probe_ollama_models(base_url: &str) -> Option<OpenAiModelsProbe> {
-    let host = strip_v1_suffix(base_url)?;
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(3))
-        .build()
-        .ok()?;
-    // 已加载集合：失败按空集（全部未加载），不影响已下载列表。
-    let loaded_names = match client
-        .get(format!("{host}/api/ps"))
-        .send()
-        .await
-        .and_then(|r| r.error_for_status())
-    {
-        Ok(resp) => resp
-            .json::<serde_json::Value>()
-            .await
-            .map(parse_ollama_ps_names)
-            .unwrap_or_default(),
-        Err(_) => Default::default(),
-    };
-    // 已下载列表：/api/tags 是必需项，失败则整个候选离线。
-    let resp = client
-        .get(format!("{host}/api/tags"))
-        .send()
-        .await
-        .and_then(|r| r.error_for_status())
-        .ok()?;
-    let tags = parse_ollama_tag_names(resp.json::<serde_json::Value>().await.ok()?);
-    (!tags.is_empty()).then_some(OpenAiModelsProbe {
-        models: tags
-            .into_iter()
-            .map(|name| {
-                let loaded = loaded_names.contains(&name);
-                OpenAiModelInfo {
-                    id: name,
-                    max_model_len: None,
-                    loaded: Some(loaded),
-                }
-            })
-            .collect(),
-    })
-}
-
-/// 探测 LM Studio：优先原生 `/api/v0/models`（带 loaded 状态），
-/// 旧版本没有该接口时回退 `/v1/models`（loaded 未知）。
-pub async fn probe_lmstudio_models(base_url: &str) -> Option<OpenAiModelsProbe> {
-    let host = strip_v1_suffix(base_url)?;
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(3))
-        .build()
-        .ok()?;
-    if let Ok(resp) = client
-        .get(format!("{host}/api/v0/models"))
-        .send()
-        .await
-        .and_then(|r| r.error_for_status())
-    {
-        if let Ok(v) = resp.json::<serde_json::Value>().await {
-            if let Some(models) = parse_lmstudio_v0_models(&v) {
-                return Some(OpenAiModelsProbe { models });
-            }
-        }
-    }
-    probe_openai_models(base_url).await
 }
 
 fn infer_context_window(preset: ModelPreset, model: Option<&str>) -> Option<u32> {
     // 模型名事实与 Engine route_limits 共用同一入口，避免页面显示 1M、实际仍按
-    // 128K 压缩。这里只保留底座与补充表都无法识别时的供应商预设兜底。
+    // 128K 压缩。底座与补充表都无法识别时按供应商预设兜底（表在 prefs::model_preset）。
     if let Some(window) = model.and_then(crate::core::model_context::resolved_context_window) {
         return Some(window);
     }
-    match preset {
-        ModelPreset::LocalVllm => Some(262_144),
-        ModelPreset::Deepseek => Some(131_072),
-        ModelPreset::Kimi => Some(262_144),
-        ModelPreset::OpenaiCompatible => Some(131_072),
-        ModelPreset::Qwen => Some(131_072),
-        ModelPreset::Doubao => Some(262_144),
-        ModelPreset::Minimax => Some(204_800),
-        ModelPreset::Glm => Some(131_072),
-        ModelPreset::Mimo => Some(1_000_000),
-        // OpenAI 官方口径：gpt-5.4-mini / gpt-5.3-codex 为 400K，其余现役旗舰 1.05M。
-        ModelPreset::Openai => match model.map(str::to_ascii_lowercase) {
-            Some(m) if m.contains("gpt-5.4-mini") || m.contains("gpt-5.3-codex") => Some(400_000),
-            _ => Some(1_050_000),
-        },
-        // Anthropic 官方口径：haiku 200K，opus/sonnet/fable 1M
-        // （claude-opus-5 由 model_context 的 PINVOU_OVERRIDES 先行覆盖，此处兜
-        // 底只承接底座不认识的命名）。
-        ModelPreset::Anthropic => match model.map(str::to_ascii_lowercase) {
-            Some(m) if m.contains("haiku") => Some(200_000),
-            _ => Some(1_000_000),
-        },
-        // Gemini 全系标称 1M。
-        ModelPreset::Gemini => Some(1_048_576),
-        // xAI 官方口径：grok-4.20 系 2M、grok-4.3 1M、grok-4.5 500K、grok-build 256K。
-        ModelPreset::Xai => match model.map(str::to_ascii_lowercase) {
-            Some(m) if m.contains("grok-4.20") => Some(2_000_000),
-            Some(m) if m.contains("grok-4.3") => Some(1_000_000),
-            Some(m) if m.contains("grok-4.5") => Some(500_000),
-            Some(m) if m.contains("grok-build") => Some(256_000),
-            _ => Some(256_000),
-        },
-    }
+    preset.context_window_fallback(model)
 }
 
 /// 推理性能相关的 6 个累计指标，统一解析、统一缺省 None。
@@ -1159,32 +939,6 @@ fn vllm_target_kind(upstream: &str) -> &'static str {
     }
     // 域名或公网 IPv6 → 云端 API
     "remote"
-}
-
-fn strip_v1_suffix(url: &str) -> Option<String> {
-    let trimmed = url.trim_end_matches('/');
-    Some(
-        trimmed
-            .strip_suffix("/v1")
-            .map(String::from)
-            .unwrap_or_else(|| trimmed.to_string()),
-    )
-}
-
-/// 模型列表探测地址：upstream 带 `/v1` 后缀时直接拼 `/models`；不带也拼 `/models`
-/// 而非补一层 `/v1`——glm `/paas/v4`、火山方舟 `/api/v3`、gemini `/v1beta/openai`
-/// 的 `/models` 端点均存在，补 `/v1` 会拼成不存在的地址永远 404。
-fn models_probe_url(upstream: &str) -> String {
-    format!("{}/models", upstream.trim_end_matches('/'))
-}
-
-/// Anthropic 官方端点用 x-api-key + anthropic-version 鉴权，不接受 Bearer。
-/// 与 app/commands/settings.rs test_model_connection 同款判定。
-fn is_anthropic_endpoint(base_url: &str) -> bool {
-    reqwest::Url::parse(base_url.trim())
-        .ok()
-        .and_then(|url| url.host_str().map(String::from))
-        .is_some_and(|host| host.eq_ignore_ascii_case("api.anthropic.com"))
 }
 
 /// 轻量探测本地 vLLM:一次 `/v1/models` 拿两样——实际 served 模型名 + `max_model_len`
@@ -1362,22 +1116,6 @@ mod tests {
         assert!(parse_prom_metric(text, "vllm:num_requests_running").is_none());
     }
 
-    #[test]
-    fn strip_v1_suffix_removes_trailing_v1() {
-        assert_eq!(
-            strip_v1_suffix("http://host:8000/v1").as_deref(),
-            Some("http://host:8000")
-        );
-        assert_eq!(
-            strip_v1_suffix("http://host:8000/v1/").as_deref(),
-            Some("http://host:8000")
-        );
-        assert_eq!(
-            strip_v1_suffix("http://host:8000").as_deref(),
-            Some("http://host:8000")
-        );
-    }
-
     #[tokio::test]
     async fn sample_all_keeps_other_fields_when_cpu_snapshot_is_none() {
         let state = MonitorState::new();
@@ -1414,66 +1152,6 @@ mod tests {
         let (id, max) = parse_models_response(json).unwrap();
         assert_eq!(id.as_deref(), Some("/model"));
         assert_eq!(max, Some(65536));
-    }
-
-    #[test]
-    fn parse_models_response_list_keeps_all_model_ids() {
-        let json: serde_json::Value = serde_json::from_str(
-            r#"{"object":"list","data":[{"id":"qwen2.5-coder:32b"},{"id":"deepseek-r1:14b","max_model_len":32768}]}"#,
-        )
-        .unwrap();
-        let models = parse_models_response_list(json).unwrap();
-        assert_eq!(models.len(), 2);
-        assert_eq!(models[0].id, "qwen2.5-coder:32b");
-        assert_eq!(models[0].max_model_len, None);
-        assert_eq!(models[1].id, "deepseek-r1:14b");
-        assert_eq!(models[1].max_model_len, Some(32768));
-    }
-
-    #[test]
-    fn ollama_ps_names_collects_loaded_models() {
-        let json: serde_json::Value = serde_json::from_str(
-            r#"{"models":[{"name":"qwen3:8b","model":"qwen3:8b","size_vram":5000000000},{"model":"deepseek-r1:14b"}]}"#,
-        )
-        .unwrap();
-        let names = parse_ollama_ps_names(json);
-        assert!(names.contains("qwen3:8b"));
-        assert!(names.contains("deepseek-r1:14b")); // 缺 name 时回退 model 字段
-        assert!(!names.contains("llama3.2:3b"));
-        // 坏形状按空集（宁全标未加载，不错标已加载）
-        assert!(parse_ollama_ps_names(serde_json::json!({})).is_empty());
-    }
-
-    #[test]
-    fn ollama_tag_names_dedupes_and_keeps_order() {
-        let json: serde_json::Value = serde_json::from_str(
-            r#"{"models":[{"name":"qwen3:8b"},{"name":"deepseek-r1:14b"},{"name":"qwen3:8b"},{"name":" "}]}"#,
-        )
-        .unwrap();
-        assert_eq!(
-            parse_ollama_tag_names(json),
-            vec!["qwen3:8b".to_string(), "deepseek-r1:14b".to_string()]
-        );
-    }
-
-    #[test]
-    fn lmstudio_v0_models_parse_loaded_state() {
-        let json: serde_json::Value = serde_json::from_str(
-            r#"{"object":"list","data":[
-                {"id":"qwen3-8b","state":"loaded"},
-                {"id":"deepseek-r1-14b","state":"not-loaded"},
-                {"id":"legacy-model"}
-            ]}"#,
-        )
-        .unwrap();
-        let models = parse_lmstudio_v0_models(&json).unwrap();
-        assert_eq!(models.len(), 3);
-        assert_eq!(models[0].loaded, Some(true));
-        assert_eq!(models[1].loaded, Some(false));
-        // 缺 state 字段 = 未知；空列表 / 坏形状返回 None，调用方回退 OpenAI 兼容探测。
-        assert_eq!(models[2].loaded, None);
-        assert!(parse_lmstudio_v0_models(&serde_json::json!({"data":[]})).is_none());
-        assert!(parse_lmstudio_v0_models(&serde_json::json!({})).is_none());
     }
 
     /// 真机冒烟(#[ignore]):对活着的本地 vLLM 跑 `probe_vllm_model_info`,确认拿到
@@ -1601,21 +1279,10 @@ vllm:request_time_per_output_token_seconds_sum{engine=\"0\",model_name=\"qwen36_
             (ModelPreset::OpenaiCompatible, "gpt-5.6-terra", 1_050_000),
             (ModelPreset::OpenaiCompatible, "gpt-5.6-luna", 1_050_000),
             (ModelPreset::OpenaiCompatible, "gpt-5.6-sol", 1_050_000),
-            // OpenAI 预设兜底：gpt-5.4-mini / gpt-5.3-codex 400K，其余 1.05M
-            (ModelPreset::Openai, "gpt-5.4-mini", 400_000),
-            (ModelPreset::Openai, "gpt-5.3-codex", 400_000),
-            (ModelPreset::Openai, "gpt-5.6-sol", 1_050_000),
-            // Anthropic：haiku 200K（底座 catalog 已知）；claude-opus-5 经
-            // model_context 的 PINVOU_OVERRIDES 覆盖底座 200K 通配兜底为 1M；
-            // 底座不认识的非 claude 命名模型走预设兜底 1M
+            // 底座 catalog 已知（haiku 200K）与 PINVOU_OVERRIDES 覆盖（opus-5 1M）
+            // 的 Anthropic 模型走 resolved_context_window，preset 兜底见 prefs 测试。
             (ModelPreset::Anthropic, "claude-haiku-4-5", 200_000),
             (ModelPreset::Anthropic, "claude-opus-5", 1_000_000),
-            (ModelPreset::Anthropic, "anthropic-future-model", 1_000_000),
-            // xAI 预设兜底：grok-4.20 系 2M、grok-4.3 1M、grok-4.5 500K、grok-build 256K
-            (ModelPreset::Xai, "grok-4.20-0309-reasoning", 2_000_000),
-            (ModelPreset::Xai, "grok-4.3", 1_000_000),
-            (ModelPreset::Xai, "grok-4.5", 500_000),
-            (ModelPreset::Xai, "grok-build-0.1", 256_000),
         ];
         for (preset, model, expected) in cases {
             assert_eq!(
@@ -1626,7 +1293,8 @@ vllm:request_time_per_output_token_seconds_sum{engine=\"0\",model_name=\"qwen36_
         }
     }
 
-    /// 推断优先级：显式 Nk 后缀 > pinvou 补充表/底座 > 预设兜底。
+    /// 推断优先级：显式 Nk 后缀 > pinvou 补充表/底座 > 预设兜底
+    /// （兜底表逐厂商覆盖见 prefs::model_preset 的 context_window_fallback 测试）。
     #[test]
     fn infer_context_window_fallback_order() {
         // 显式后缀优先于一切（含底座 catalog 里的同名模型）
@@ -1639,72 +1307,6 @@ vllm:request_time_per_output_token_seconds_sum{engine=\"0\",model_name=\"qwen36_
             infer_context_window(ModelPreset::Deepseek, Some("my-custom-finetune")),
             Some(131_072)
         );
-        assert_eq!(
-            infer_context_window(ModelPreset::Minimax, Some("minimax-future-x")),
-            Some(204_800)
-        );
-        assert_eq!(
-            infer_context_window(ModelPreset::Mimo, Some("mimo-future-x")),
-            Some(1_000_000)
-        );
-        // 无模型名 → 预设兜底
         assert_eq!(infer_context_window(ModelPreset::Kimi, None), Some(262_144));
-        // Openai/Anthropic/Xai 臂的无模型名缺省
-        assert_eq!(
-            infer_context_window(ModelPreset::Openai, None),
-            Some(1_050_000)
-        );
-        assert_eq!(
-            infer_context_window(ModelPreset::Anthropic, None),
-            Some(1_000_000)
-        );
-        assert_eq!(infer_context_window(ModelPreset::Xai, None), Some(256_000));
-        // Xai 臂不认识的未来型号 → 缺省 256K
-        assert_eq!(
-            infer_context_window(ModelPreset::Xai, Some("grok-future-x")),
-            Some(256_000)
-        );
-    }
-
-    /// 模型探测地址：带 `/v1` 结尾保持既有行为；不带时不补 `/v1`，直接拼 `/models`
-    /// （glm `/paas/v4`、火山方舟 `/api/v3`、gemini `/v1beta/openai` 的 `/models` 均存在）。
-    #[test]
-    fn models_probe_url_appends_models_without_extra_v1() {
-        assert_eq!(
-            models_probe_url("http://127.0.0.1:8000/v1"),
-            "http://127.0.0.1:8000/v1/models"
-        );
-        assert_eq!(
-            models_probe_url("http://127.0.0.1:8000/v1/"),
-            "http://127.0.0.1:8000/v1/models"
-        );
-        assert_eq!(
-            models_probe_url("https://open.bigmodel.cn/api/paas/v4"),
-            "https://open.bigmodel.cn/api/paas/v4/models"
-        );
-        assert_eq!(
-            models_probe_url("https://ark.cn-beijing.volces.com/api/v3"),
-            "https://ark.cn-beijing.volces.com/api/v3/models"
-        );
-        assert_eq!(
-            models_probe_url("https://generativelanguage.googleapis.com/v1beta/openai"),
-            "https://generativelanguage.googleapis.com/v1beta/openai/models"
-        );
-        assert_eq!(
-            models_probe_url("https://api.anthropic.com"),
-            "https://api.anthropic.com/models"
-        );
-    }
-
-    /// Anthropic 官方端点判定：仅 api.anthropic.com 主机走 x-api-key 鉴权。
-    #[test]
-    fn is_anthropic_endpoint_matches_only_official_host() {
-        assert!(is_anthropic_endpoint("https://api.anthropic.com"));
-        assert!(is_anthropic_endpoint("https://api.anthropic.com/v1"));
-        assert!(is_anthropic_endpoint("https://API.ANTHROPIC.COM"));
-        assert!(!is_anthropic_endpoint("https://api.openai.com/v1"));
-        assert!(!is_anthropic_endpoint("https://anthropic.example.com"));
-        assert!(!is_anthropic_endpoint("http://127.0.0.1:8000/v1"));
-        assert!(!is_anthropic_endpoint("not a url"));
     }
 }
