+++ pinvou3-app/src/shared/i18n.js
@@ -1167,147 +1167,7 @@ const dict = {
     };
     // ja 漏翻的 key 回落英文，避免新增文案后显示 undefined
     dict.ja = { ...dict.en, ...dict.ja };
-    Object.assign(dict.zh.uiSettingsDetail, {
-      addProvider:provider=>`添加 ${provider}`, editProvider:provider=>`编辑 ${provider}`,
-      autoDetectLocalModel:'自动检测本地模型',
-      codingPlanTestUnavailable:'当前厂商接口暂时无法完成测试，但不影响保存配置',
-      connectionMessages:{ ok:'连接成功，服务可用', auth_invalid:'API Key 无效，请检查后重新填写', auth_forbidden:'当前 API Key 没有访问权限', rate_limited:'请求过于频繁或额度不足，请稍后再试', endpoint_not_found:'接口地址不可用', method_not_allowed:'接口不支持测试请求', http_error:'连接失败，请检查配置后重试', unknown:'连接失败，请稍后重试' },
-      catalogSections:{ coding_plan:'Coding Plan', official_api:'官方 API', custom:'自定义兼容接口' },
-      providerCatalog:{},
-    });
-    Object.assign(dict.en.uiSettingsDetail, {
-      saving:'Saving', localDetectionTargets:'Detect vLLM, Ollama, and LM Studio',
-      redetect:'Detect again', detect:'Detect', noRunningLocalModel:'No running local model detected', modelNotLoadedTag:'Not loaded', modelNotLoadedHint:'Not in memory yet; it will load automatically on first use',
-      add:'Add', manualLocalModel:'Add a local model manually',
-      manualLocalModelDesc:'Enter the API URL and model ID', cloudModels:'Cloud models',
-      localModels:'Local models', toolCalling:'Tool calling', apiKeyRequired:'Requires API Key',
-      testBeforeSave:'Test whether the service is available before saving',
-      addProvider:provider=>`Add ${provider}`, editProvider:provider=>`Edit ${provider}`,
-      autoDetectLocalModel:'Automatically detect local models',
-      codingPlanTestUnavailable:'This provider endpoint cannot be tested right now, but you can still save the configuration',
-      connectionMessages:{ ok:'Connection successful; service available', auth_invalid:'Invalid API Key. Check it and try again', auth_forbidden:'This API Key does not have access', rate_limited:'Too many requests or insufficient quota. Try again later', endpoint_not_found:'Endpoint unavailable', method_not_allowed:'The endpoint does not support test requests', http_error:'Connection failed. Check the configuration and retry', unknown:'Connection failed. Try again later' },
-      catalogSections:{ coding_plan:'Coding Plan', official_api:'Official APIs', custom:'Custom compatible endpoints' },
-      providerCatalog:{
-        local:{title:'Local Model',desc:'Default model for the local service'}, glm_coding_plan:{title:'Zhipu Coding Plan / GLM Coding Plan',configTitle:'Zhipu Coding Plan',desc:'Dedicated endpoint for coding and agent tasks'},
-        tencent_coding_plan:{title:'Tencent Cloud Coding Plan',configTitle:'Tencent Cloud Coding Plan',desc:'Tencent Cloud coding-plan endpoint'},
-        kimi_coding_plan:{title:'Kimi Coding Plan',configTitle:'Kimi Coding Plan',desc:'Dedicated endpoint for Kimi coding tasks'},
-        deepseek:{title:'DeepSeek',configTitle:'DeepSeek',desc:'Official DeepSeek API'}, kimi:{title:'Kimi China',configTitle:'Kimi',desc:'Official Moonshot API'},
-        glm:{title:'Zhipu Open Platform / GLM API',configTitle:'GLM API',desc:'Standard Zhipu Open Platform API'},
-        minimax:{title:'MiniMax China',configTitle:'MiniMax',desc:'Official MiniMax API'}, mimo:{title:'MiMo',desc:'Official Xiaomi MiMo API'},
-        qwen:{title:'Qwen',desc:'Alibaba Cloud DashScope-compatible API'}, doubao:{title:'Doubao',desc:'Official Volcengine Ark API'},
-        glm_coding_plan_global:{title:'Zhipu Coding Plan Global / GLM Coding Plan Global',configTitle:'Zhipu Coding Plan Global',desc:'Dedicated z.ai endpoint for coding and agent tasks'},
-        kimi_global:{title:'Kimi Global',configTitle:'Kimi Global',desc:'Moonshot international API'},
-        glm_global:{title:'Zhipu GLM API (z.ai)',configTitle:'GLM API (z.ai)',desc:'Zhipu international z.ai API'},
-        minimax_global:{title:'MiniMax Global',configTitle:'MiniMax Global',desc:'MiniMax international API (keys are region-locked)'},
-        qwen_token_plan:{title:'Qwen Token Plan',configTitle:'Qwen Token Plan',desc:'Dedicated gateway for Alibaba Token Plan subscriptions'},
-        qwen_global:{title:'Qwen International',configTitle:'Qwen International',desc:'Alibaba Cloud Model Studio international API'},
-        openai:{title:'OpenAI',configTitle:'OpenAI',desc:'Official OpenAI API'},
-        anthropic:{title:'Anthropic Claude',configTitle:'Anthropic Claude',desc:'Official Anthropic API (native Messages protocol)'},
-        gemini:{title:'Google Gemini',configTitle:'Google Gemini',desc:'Gemini API (OpenAI-compatible endpoint)'},
-        xai:{title:'xAI Grok',configTitle:'xAI Grok',desc:'Official xAI API'},
-        openai_compatible:{title:'OpenAI Compatible',desc:'Custom OpenAI-compatible endpoint'},
-      },
-    });
-    Object.assign(dict.ja.uiSettingsDetail, {
-      saving:'保存中', localDetectionTargets:'vLLM、Ollama、LM Studio を検出',
-      redetect:'再検出', detect:'検出', noRunningLocalModel:'実行中のローカルモデルが見つかりません', modelNotLoadedTag:'未読み込み', modelNotLoadedHint:'メモリ未読み込み。初回使用時に自動でロードされます',
-      add:'追加', manualLocalModel:'ローカルモデルを手動追加',
-      manualLocalModelDesc:'API URL とモデル ID を入力', cloudModels:'クラウドモデル',
-      localModels:'ローカルモデル', toolCalling:'ツール呼び出し', apiKeyRequired:'API Key が必要',
-      testBeforeSave:'保存前にサービスが利用可能かテストできます',
-      addProvider:provider=>`${provider} を追加`, editProvider:provider=>`${provider} を編集`,
-      autoDetectLocalModel:'ローカルモデルを自動検出',
-      codingPlanTestUnavailable:'現在このプロバイダーの接続テストはできませんが、設定は保存できます',
-      connectionMessages:{ ok:'接続に成功しました', auth_invalid:'API Key が無効です。確認して再入力してください', auth_forbidden:'この API Key にはアクセス権がありません', rate_limited:'リクエストが多すぎるか、割り当てが不足しています。後でもう一度お試しください', endpoint_not_found:'エンドポイントを利用できません', method_not_allowed:'エンドポイントはテストリクエストに対応していません', http_error:'接続に失敗しました。設定を確認して再試行してください', unknown:'接続に失敗しました。後でもう一度お試しください' },
-      catalogSections:{ coding_plan:'Coding Plan', official_api:'公式 API', custom:'カスタム互換エンドポイント' },
-      providerCatalog:{
-        local:{title:'ローカルモデル',desc:'ローカルサービスのデフォルトモデル'}, glm_coding_plan:{title:'Zhipu Coding Plan / GLM Coding Plan',configTitle:'Zhipu Coding Plan',desc:'コーディングと Agent 向けの専用エンドポイント'},
-        tencent_coding_plan:{title:'Tencent Cloud Coding Plan',configTitle:'Tencent Cloud Coding Plan',desc:'Tencent Cloud のコーディングプラン用エンドポイント'},
-        kimi_coding_plan:{title:'Kimi Coding Plan',configTitle:'Kimi Coding Plan',desc:'Kimi のコーディング向け専用エンドポイント'},
-        deepseek:{title:'DeepSeek',configTitle:'DeepSeek',desc:'DeepSeek 公式 API'}, kimi:{title:'Kimi China',configTitle:'Kimi',desc:'Moonshot 公式 API'},
-        glm:{title:'Zhipu Open Platform / GLM API',configTitle:'GLM API',desc:'Zhipu Open Platform の標準 API'},
-        minimax:{title:'MiniMax China',configTitle:'MiniMax',desc:'MiniMax 公式 API'}, mimo:{title:'MiMo',desc:'Xiaomi MiMo 公式 API'},
-        qwen:{title:'Qwen',desc:'Alibaba Cloud DashScope 互換 API'}, doubao:{title:'Doubao',desc:'Volcengine Ark 公式 API'},
-        glm_coding_plan_global:{title:'Zhipu Coding Plan Global / GLM Coding Plan Global',configTitle:'Zhipu Coding Plan Global',desc:'z.ai のコーディング・Agent 向け専用エンドポイント'},
-        kimi_global:{title:'Kimi Global',configTitle:'Kimi Global',desc:'Moonshot 国際版 API'},
-        glm_global:{title:'Zhipu GLM API (z.ai)',configTitle:'GLM API (z.ai)',desc:'Zhipu 国際版 z.ai API'},
-        minimax_global:{title:'MiniMax Global',configTitle:'MiniMax Global',desc:'MiniMax 国際版 API（キーは国内版と共通ではありません）'},
-        qwen_token_plan:{title:'Qwen Token Plan',configTitle:'Qwen Token Plan',desc:'Alibaba Token Plan サブスクリプション専用ゲートウェイ'},
-        qwen_global:{title:'Qwen International',configTitle:'Qwen International',desc:'Alibaba Cloud Model Studio 国際版 API'},
-        openai:{title:'OpenAI',configTitle:'OpenAI',desc:'OpenAI 公式 API'},
-        anthropic:{title:'Anthropic Claude',configTitle:'Anthropic Claude',desc:'Anthropic 公式 API（Messages ネイティブプロトコル）'},
-        gemini:{title:'Google Gemini',configTitle:'Google Gemini',desc:'Gemini API（OpenAI 互換エンドポイント）'},
-        xai:{title:'xAI Grok',configTitle:'xAI Grok',desc:'xAI 公式 API'},
-        openai_compatible:{title:'OpenAI Compatible',desc:'カスタム OpenAI 互換エンドポイント'},
-      },
-    });
-    Object.assign(dict.en.uiSettingsDetail.modelDescriptions, {
-      '本地服务默认模型':'Default model for the local service',
-      '旗舰编码模型':'Flagship coding model', '高性能编码模型':'High-performance coding model',
-      '日常编码模型':'Everyday coding model', 'Coding Plan 自动模型':'Coding Plan automatic model',
-      '标准编码模型':'Standard coding model', 'K3 256K 上下文模型':'K3 with 256K context',
-      'K3 长上下文模型':'K3 long-context model', '高速编码模型':'High-speed coding model',
-      '旗舰推理':'Flagship reasoning model',
-      'K3 256K 上下文，价格更低':'K3 with 256K context at a lower price',
-      '官方已转 Legacy，兼容保留':'Officially legacy; kept for compatibility',
-      '官方已转 Legacy，兼容高速':'Officially legacy; high-speed compatibility',
-      '最新旗舰':'Latest flagship', '上代旗舰推理':'Previous-generation flagship reasoning',
-      '均衡性价比':'Balanced and cost-effective', '快速高性价比':'Fast and cost-effective',
-      '正式旗舰，夜间五折':'GA flagship, 50% off at night',
-      '2.4T 旗舰预览，Token Plan 专属，预览结束将下线或替换':'2.4T flagship preview, Token Plan only; will be retired or replaced after the preview',
-      '旗舰推理与编码':'Flagship reasoning and coding', '均衡智能与成本':'Balanced intelligence and cost',
-      '低成本高并发':'Low cost, high volume', '上代旗舰':'Previous-generation flagship',
-      '快速经济':'Fast and economical', '最强旗舰，长程 Agent':'Most capable flagship for long-running agents',
-      '复杂 Agent 编码':'Complex agentic coding', '速度与智能均衡':'Balanced speed and intelligence',
-      '最快，接近旗舰':'Fastest, near-frontier', '最新 Flash，均衡高性价比':'Latest Flash, balanced and cost-effective',
-      '均衡':'Balanced', '旗舰推理（预览）':'Flagship reasoning (preview)',
-      '4.20 推理':'4.20 reasoning', '4.20 非推理':'4.20 non-reasoning',
-      '旗舰编码与 Agent':'Flagship coding and agent model',
-      '通用推理，默认推荐':'General-purpose reasoning, recommended default', '代码 Agent':'Coding agent',
-    });
-    Object.assign(dict.ja.uiSettingsDetail.modelDescriptions, {
-      '本地服务默认模型':'ローカルサービスのデフォルトモデル',
-      '旗舰编码模型':'フラッグシップコーディングモデル', '高性能编码模型':'高性能コーディングモデル',
-      '日常编码模型':'日常向けコーディングモデル', 'Coding Plan 自动模型':'Coding Plan 自動モデル',
-      '标准编码模型':'標準コーディングモデル', 'K3 256K 上下文模型':'K3 256K コンテキストモデル',
-      'K3 长上下文模型':'K3 長コンテキストモデル', '高速编码模型':'高速コーディングモデル',
-      '旗舰推理':'フラッグシップ推論モデル',
-      'K3 256K 上下文，价格更低':'K3 256K コンテキスト、より低価格',
-      '官方已转 Legacy，兼容保留':'公式レガシー化済み。互換性のため保持',
-      '官方已转 Legacy，兼容高速':'公式レガシー化済み。高速互換',
-      '最新旗舰':'最新フラッグシップ', '上代旗舰推理':'前世代フラッグシップ推論',
-      '均衡性价比':'バランス型で高コストパフォーマンス', '快速高性价比':'高速で高コストパフォーマンス',
-      '正式旗舰，夜间五折':'正式版フラッグシップ、夜間半額',
-      '2.4T 旗舰预览，Token Plan 专属，预览结束将下线或替换':'2.4T フラッグシッププレビュー、Token Plan 専用。プレビュー終了後は廃止または差し替え',
-      '旗舰推理与编码':'フラッグシップ推論とコーディング', '均衡智能与成本':'知能とコストのバランス',
-      '低成本高并发':'低コスト・高スループット', '上代旗舰':'前世代フラッグシップ',
-      '快速经济':'高速で経済的', '最强旗舰，长程 Agent':'最強フラッグシップ、長時間エージェント向け',
-      '复杂 Agent 编码':'複雑なエージェントコーディング', '速度与智能均衡':'速度と知能のバランス',
-      '最快，接近旗舰':'最速、フラッグシップに近い', '最新 Flash，均衡高性价比':'最新 Flash、バランス型で高コストパフォーマンス',
-      '均衡':'バランス型', '旗舰推理（预览）':'フラッグシップ推論（プレビュー）',
-      '4.20 推理':'4.20 推論', '4.20 非推理':'4.20 非推論',
-      '旗舰编码与 Agent':'フラッグシップコーディング・Agent モデル',
-      '通用推理，默认推荐':'汎用推論、デフォルト推奨', '代码 Agent':'コーディングエージェント',
-    });
-    // 自定义项标题/描述的逐供应商口径（与 i18n 化前硬编码目录一致）；未列出的供应商回退 customModelTitle。
-    dict.zh.uiSettingsDetail.customCodingPlanDesc = '手动填写 Coding Plan 模型 ID';
-    dict.en.uiSettingsDetail.customCodingPlanDesc = 'Enter a Coding Plan model ID manually';
-    dict.ja.uiSettingsDetail.customCodingPlanDesc = 'Coding Plan モデル ID を手動入力';
-    dict.zh.uiSettingsDetail.customModelTitles = {
-      glm:'自定义 GLM 模型', qwen:'自定义通义模型',
-      openai_compatible:'自定义兼容模型', glm_coding_plan:'自定义 GLM Coding Plan 模型',
-      tencent_coding_plan:'自定义腾讯云 Coding Plan 模型', kimi_coding_plan:'自定义 Kimi Coding Plan 模型',
-    };
-    dict.en.uiSettingsDetail.customModelTitles = {
-      glm:'Custom GLM model', qwen:'Custom Qwen model',
-      openai_compatible:'Custom compatible model', glm_coding_plan:'Custom GLM Coding Plan model',
-      tencent_coding_plan:'Custom Tencent Cloud Coding Plan model', kimi_coding_plan:'Custom Kimi Coding Plan model',
-    };
-    dict.ja.uiSettingsDetail.customModelTitles = {
-      glm:'カスタム GLM モデル', qwen:'カスタム Qwen モデル',
-      openai_compatible:'カスタム互換モデル', glm_coding_plan:'カスタム GLM Coding Plan モデル',
-      tencent_coding_plan:'カスタム Tencent Cloud Coding Plan モデル', kimi_coding_plan:'カスタム Kimi Coding Plan モデル',
-    };
+    // 设置详情(uiSettingsDetail)的补丁块已归位 features/settings/settings-i18n.js,随模型目录一同维护。
     dict.zh.uiToolDetails.flow = {
       incomplete:name=>`${name}接入未完成`, connected:name=>`已连接${name}`, connecting:name=>`正在接入${name}`,
       cancel:'取消', extracting:pct=>`解压中 ${pct}%`, elapsed:seconds=>`已 ${seconds}s`, installStarting:'在线安装：正在开始…', browserOpened:'已打开浏览器登录页',
